requirements.txt
requests>=2.28
beautifulsoup4>=4.12
Jinja2>=3.1
PyGithub>=1.78
google-genai>=0.3.0
pytest>=8.2